Refactor and simplify types in MatchingComparisonOutput
