+++ packages/core/tests/helpers/summariseMatching.ts
@@ -1,7 +1,7 @@
 import ExceptionCode from "@moj-bichard7-developers/bichard7-next-data/dist/types/ExceptionCode"
 
 import type { AnnotatedHearingOutcome } from "../../types/AnnotatedHearingOutcome"
-import type { CourtResultMatchingSummary } from "../types/MatchingComparisonOutput"
+import type CourtResultMatchingSummary from "../types/CourtResultMatchingSummary"
 
 import getOffenceCode from "../../lib/offences/getOffenceCode"
